log_audio immediately updates "audios/" in neptune & try gl_dtype argument to generate to see if quality improves

+++ mimikit/freqnet/base.py
@@ -2,6 +2,7 @@ import torch
 from pytorch_lightning import LightningModule, LightningDataModule
 from torch.utils.data import DataLoader
 import librosa
+import numpy as np
 from abc import ABC
 
 from ..data import DataObject, FeatureProxy, HOP_LENGTH
@@ -209,7 +210,7 @@ class FreqNetModel(MMKHooks,
     def targets_shifts_and_lengths(self, input_length):
         raise NotImplementedError("subclasses of `FreqNetModel` have to implement `targets_shifts_and_lengths`")
 
-    def generate(self, prompt, n_steps, hop_length=HOP_LENGTH, time_domain=True):
+    def generate(self, prompt, n_steps, hop_length=HOP_LENGTH, time_domain=True, gl_dtype=np.float32):
         was_training = self.training
         self.eval()
         initial_device = self.device
@@ -226,8 +227,8 @@ class FreqNetModel(MMKHooks,
                 generated = torch.cat((generated, out[:, output_slice]), dim=1)
         if time_domain:
             generated = generated.transpose(1, 2).squeeze()
-            generated = librosa.griffinlim(generated.cpu().numpy(), hop_length=hop_length, n_iter=64)
-            generated = torch.from_numpy(generated)
+            generated = librosa.griffinlim(generated.cpu().numpy().astype(gl_dtype), hop_length=hop_length, n_iter=64)
+            generated = torch.from_numpy(generated.astype(np.float32))
         self.to(initial_device)
         self.train() if was_training else None
         return generated.unsqueeze(0)
